+++ symbolic-symcache/src/raw/mod.rs
@@ -2,6 +2,7 @@
 //!
 
 pub(crate) mod v7;
+pub(crate) mod v9;
 
 use watto::Pod;